refactor ResourceLoaderOverrides.js
